Combine tests/functional/test_vcs.py with other test modules.
